adds form validator for new and edit game pages

+++ src/apis/views.py
@@ -1,6 +1,14 @@
+from re import A
+
 from app import app
 from flask import flash, render_template, request, send_from_directory
-from services import AuthService, GameService, ImageService, RedirectService
+from services import (
+    AuthService,
+    GameService,
+    GameValidatorService,
+    ImageService,
+    RedirectService,
+)
 
 NEXT_PAGE = "next"
 
@@ -67,9 +75,12 @@ def new():
     if not AuthService().is_authenticated():
         return RedirectService().to_login(next_page="new")
 
+    form = GameValidatorService()
+
     return render_template(
         "new_game.html",
         title="New Game",
+        form=form,
     )
 
 
@@ -79,12 +90,19 @@ def edit(id):
         return RedirectService().to_login(next_page="edit")
 
     game = GameService().get(id=id)
+
+    form = GameValidatorService()
+    form.name.data = game.name
+    form.category.data = game.category
+    form.platform.data = game.platform
+
     cover = ImageService().get_image(id=id)
 
     return render_template(
         "edit_game.html",
         title="Edit Game",
-        game=game,
+        id=game.id,
+        form=form,
         cover=cover,
     )
 
@@ -108,9 +126,14 @@ def delete(id):
     ],
 )
 def create():
-    name = request.form["name"]
-    category = request.form["category"]
-    platform = request.form["platform"]
+    form = GameValidatorService(request.form)
+
+    if not form.validate_on_submit():
+        return RedirectService().to_page("new")
+
+    name = form.name.data
+    category = form.category.data
+    platform = form.platform.data
 
     if GameService().get(name=name):
         flash("O jogo já está registrado!")
@@ -135,20 +158,23 @@ def create():
     ],
 )
 def update():
-    id = request.form["id"]
+    form = GameValidatorService(request.form)
 
-    game = GameService().update(
-        id=id,
-        name=request.form["name"],
-        category=request.form["category"],
-        platform=request.form["platform"],
-    )
+    if form.validate_on_submit():
+        id = request.form["id"]
 
-    ImageService().delete_image(id=game.id)
+        game = GameService().update(
+            id=id,
+            name=form.name.data,
+            category=form.category.data,
+            platform=form.platform.data,
+        )
 
-    picture = request.files["file"]
-    picture_name = ImageService().new_name(id=game.id)
-    picture.save(picture_name)
+        ImageService().delete_image(id=game.id)
+
+        picture = request.files["file"]
+        picture_name = ImageService().new_name(id=game.id)
+        picture.save(picture_name)
 
     return RedirectService().to_index()
